NimSG improvements

+++ spec/nim/tests/tbuffered_struct.nim
@@ -12,12 +12,12 @@ suite "Kaitai Struct Compiler Test Suite":
   test "Nim: BufferedStruct":
     let r = BufferedStruct.fromFile("src/buffered_struct.bin")
 
-    check(r.len1 == 16)
-    check(r.block1.number1 == 66)
-    check(r.block1.number2 == 67)
-    check(r.len2 == 8)
-    check(r.block2.number1 == 68)
-    check(r.block2.number2 == 69)
-    check(r.finisher == 238)
+    check(r.len1 == uint32(16))
+    check(r.block1.number1 == uint32(66))
+    check(r.block1.number2 == uint32(67))
+    check(r.len2 == uint32(8))
+    check(r.block2.number1 == uint32(68))
+    check(r.block2.number2 == uint32(69))
+    check(r.finisher == uint32(238))
 
 close(outputFormatter)
